Struct ipld_hamt::BytesKey

pub struct BytesKey(pub Vec<u8, Global>);

Key type to be used to serialize as byte string instead of a u8 array. This type is used as a default for the Hamt as this is the only allowed type with the go implementation.